dockerized

+++ 01_rnadisease2ensg.py
@@ -3,25 +3,33 @@ import csv
 from modules import rnadisease
 import pandas as pd
 import os
+import yaml
 
-parser = argparse.ArgumentParser(description="This script collects genes from RNAdisease based on a DOID (Disease Ontology ID) and annotation score, and writes it to a tsv file (./results).")
-parser.add_argument("DOID", type=str, help="DOID to search for")
-parser.add_argument("score",type=float, help="Annotation score to filter the search hits")
 
-args = parser.parse_args()
-
-def main():
+def main(config):
     """
     
     """
+    with open(f'./{config}','r') as yml:
+        config = yaml.safe_load(yml)
+    doid = config["DISEASE_ONTOLOGY_ID"]
+    score = config["RNA_DISEASE_SCORE"]
+    score = float(score)
     if not os.path.exists(f"./results"):
         os.mkdir(f"./results")
-    rnadisease.download_file()
-    rnadisease.convert_xlsx_to_csv()
-    rnadisease.rnadisease2genes(args.DOID,args.score)
-    ensg_list = rnadisease.rna2ensg()
-    df = pd.DataFrame(ensg_list)
-    df.to_csv(f"./results/rnadisease_genes.tsv",sep="\t",index=False)
+    if not os.path.exists(f"./results/rnadisease/RNADiseasev4.0_RNA-disease_experiment_all.xlsx"):
+        rnadisease.download_file()
+    if not os.path.exists(f"./results/rnadisease/rnadisease.csv"):
+        rnadisease.convert_xlsx_to_csv()
+    if not os.path.exists(f"./results/rnadisease/pre_rnadisease.tsv"):
+        rnadisease.rnadisease2genes(doid,score)
+    if not os.path.exists(f"./results/rnadisease_genes.tsv"):
+        ensg_list = rnadisease.rna2ensg()
+        df = pd.DataFrame(ensg_list)
+        df.to_csv(f"./results/rnadisease_genes.tsv",sep="\t",index=False)
+        return print("01_rnadisease2ensg has been completed")
+    else:
+        return print("The file already exists.")
 
 if __name__ == "__main__":
-    main()
+    main("./config.yml")
