+++ 03_disgenet2ensg.py
@@ -3,24 +3,30 @@ import os
 import csv
 from modules import disgenet
 import pandas as pd
+import yaml
 
-parser = argparse.ArgumentParser(description="This script collects genes associated with disease from DisGeNET based on a ncit disease term, and writes it to a tsv file (./results).")
-parser.add_argument("linked_data_Disease_ID", help="linked life data disease ID to search for")
-parser.add_argument("score", type=float,help="minimum score between 0 and 1 to filter the genes")
-parser.add_argument("--limit", help="limit of the number of evidence pmids for each gene", default="15")
 
-args = parser.parse_args()
-
-def main():
+def main(config):
     """
     
     """
+    with open(f'./{config}','r') as yml:
+        config = yaml.safe_load(yml)
     if not os.path.exists(f"./results"):
         os.mkdir(f"./results")
-    ensg_list = disgenet.disgenet2list(args.linked_data_Disease_ID,args.score,args.limit)
-    print(ensg_list)
-    df = pd.DataFrame(ensg_list)
-    df.to_csv(f"./results/disgenet_genes.tsv",sep="\t",index=False)
+    if not os.path.exists(f"./results/disgenet_genes.tsv"):
+        ncit = config["NCIT_DISEASE_TERM"]
+        score = config["DISGENET_SCORE"]
+        score = float(score)
+        limit = config["DISGENET_PUBLICATION_LIMIT"]
+        limit = int(limit)
+        ensg_list = disgenet.disgenet2list(ncit,score,limit)
+        print(ensg_list)
+        df = pd.DataFrame(ensg_list)
+        df.to_csv(f"./results/disgenet_genes.tsv",sep="\t",index=False)
+        return print("03_disgenet2ensg has been completed")
+    else:
+        return print("The file already exists.")
 
 if __name__ == "__main__":
-    main()
+    main("./config.yml")
