+++ 04_pubchem2ensg.py
@@ -4,27 +4,32 @@ import csv
 from modules import pubchem
 import pandas as pd
 import argparse
+import yaml
 
-parser = argparse.ArgumentParser(description="This script collects genes associated with disease from pubchem based on a pubchem disease ID, and writes it to a tsv file (./results).")
-parser.add_argument("pubchem_diseaseID", help="pubchem disease ID (ncit disease term) to search for")
-args = parser.parse_args()
 
-def main():
+def main(config):
     """
     
     """
+    with open(f'./{config}','r') as yml:
+        config = yaml.safe_load(yml)
     if not os.path.exists(f"./results"):
         os.mkdir(f"./results")
-    pubchem_list = pubchem.pubchem2list(args.pubchem_diseaseID)
-    print(pubchem_list)
-    pubchem_list2 = []
-    for i in pubchem_list:
-        if i["PMID_PMCID"] != "":
-            pubchem_list2.append(i)
-        else:
-            pass
-    df = pd.DataFrame(pubchem_list2)
-    df.to_csv(f"./results/pubchem_genes.tsv",sep="\t",index=False)
+    pubchem_diseaseid = config["PUBCHEM_DISEASE_ID"]
+    if not os.path.exists(f"./results/pubchem_genes.tsv"):
+        pubchem_list = pubchem.pubchem2list(pubchem_diseaseid)
+        print(pubchem_list)
+        pubchem_list2 = []
+        for i in pubchem_list:
+            if i["PMID_PMCID"] != "":
+                pubchem_list2.append(i)
+            else:
+                pass
+        df = pd.DataFrame(pubchem_list2)
+        df.to_csv(f"./results/pubchem_genes.tsv",sep="\t",index=False)
+        return print("03_pubchem2ensg has been completed")
+    else:
+        return print("The file already exists.")
 
 if __name__ == "__main__":
-    main()
+    main("./config.yml")
